requirements.txt
langchain
langchain-groq
langchain_huggingface
langchain-astradb
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental
langchain-tavily
sentence-transformers
faiss-cpu
tiktoken
chromadb
pinecone
sentence-transformers
pypdf
python-dotenv
pandas
numpy
pymupdf
python-docx
docx2txt
unstructured[docx]
openpyxl
msoffcrypto-tool
jq
matplotlib
seaborn
rank_bm25
wikipedia
arxiv
langgraph
langgraph-cli[inmem]
